tools/requirements.txt
PyYAML>=5.0.0
PyGithub>=1.53
python-dotenv==0.14
